requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
Babel==2.9.1
beautifulsoup4==4.12.3
Brotli==1.1.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudscraper==1.2.71
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.9.0.post1
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.145.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflate64==1.0.1
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.31
langchain-openai==0.2.10
langchain-text-splitters==0.3.5
langsmith==0.2.11
lxml==5.3.0
magic-filter==1.0.12
marshmallow==3.25.1
multidict==6.1.0
multivolumefile==0.2.3
mutagen==1.47.0
mypy-extensions==1.0.0
numpy==1.26.4
odfpy==1.4.1
openai==1.57.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.47.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.1
psycopg2-binary==2.9.10
py7zr==0.22.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.3
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.0.0
pyparsing==3.2.1
PyPDF2==3.0.1
pyppmd==1.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzstd==0.16.2
rarfile==4.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
tenacity==9.0.0
texttable==1.7.0
tiktoken==0.7.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
XlsxWriter==3.2.0
yarl==1.18.3
yt-dlp==2025.1.15
